Reduce the tensor module's door to a door
The value side moves out of tensor/mod.rs, which is now mod declarations and
re-exports only: broadcast.rs takes the right-aligned axis pairing and the two
functions built on it, and the tensor value takes the rest.

The value file is value.rs rather than tensor.rs. A tensor/tensor.rs trips
clippy::module_inception, and "value" is the domain's own word for what a
tensor is.

broadcast.rs is the module with the fewest dependencies in the tree: ndarray
and the error type, no dtype, shape, or tensor. align_axes and
broadcast_elementwise become pub(super) so the value side and the type-level
rules can each reach what they need. broadcast_shape keeps its public path
through the re-export.

Pure movement. The broadcasting tests exercise the arithmetic operators rather
than broadcast_shape directly, so they stay with the value; broadcast.rs has
no tests of its own.
